mrrc::field_collection

Trait FieldCollection

pub trait FieldCollection {
    // Required methods
    fn add_field_to_collection(&mut self, field: Field, name: &str);
    fn get_collection(&self, name: &str) -> Option<&[Field]>;
}

A trait for managing a collection of MARC fields.

This trait defines the operations needed to manage field collections in a consistent way across different record types. Each implementation manages a specific Vec<Field> collection.

Examples

impl FieldCollection for AuthorityRecord {
    fn add_field_to_collection(&mut self, field: Field, name: &str) {
        match name {
            "see_from" => self.tracings_see_from.push(field),
            "see_also" => self.tracings_see_also.push(field),
            _ => {}
        }
    }

    fn get_collection(&self, name: &str) -> Option<&[Field]> {
        match name {
            "see_from" => Some(&self.tracings_see_from),
            "see_also" => Some(&self.tracings_see_also),
            _ => None
        }
    }
}

Required Methods

fn add_field_to_collection(&mut self, field: Field, name: &str)

Add a field to a named collection.

fn get_collection(&self, name: &str) -> Option<&[Field]>

Get a reference to a named collection.