mrrc::leader

Struct Leader

pub struct Leader {

    pub record_length: u32,
    pub record_status: char,
    pub record_type: char,
    pub bibliographic_level: char,
    pub control_record_type: char,
    pub character_coding: char,
    pub indicator_count: u8,
    pub subfield_code_count: u8,
    pub data_base_address: u32,
    pub encoding_level: char,
    pub cataloging_form: char,
    pub multipart_level: char,
    pub reserved: String,

}

MARC Leader - 24 bytes at the start of every MARC record.

Contains metadata about the record structure and content. All MARC records must begin with exactly 24 bytes of leader information.

Fields
record_length: u32

Record length (5 digits) - positions 0-4

record_status: char

Record status (1 char) - position 5

record_type: char

Type of record (1 char) - position 6

bibliographic_level: char

Bibliographic level (1 char) - position 7

control_record_type: char

Type of control record (1 char) - position 8

character_coding: char

Character coding scheme (1 char) - position 9

indicator_count: u8

Indicator count (1 digit) - position 10 (usually 2)

subfield_code_count: u8

Subfield code count (1 digit) - position 11 (usually 2)

data_base_address: u32

Base address of data (5 digits) - positions 12-16

encoding_level: char

Encoding level (1 char) - position 17

cataloging_form: char

Cataloging form (1 char) - position 18

multipart_level: char

Multipart resource record level (1 char) - position 19

reserved: String

Reserved (4 chars) - positions 20-23

Implementations

impl Leader

pub fn from_bytes(bytes: &[u8]) -> Result<Self>

Parse a leader from 24 bytes

Errors

Returns an error if the bytes are invalid or too short.

pub fn as_bytes(&self) -> Result<Vec<u8>>

Serialize leader to 24-byte array

Errors

Returns an error if the leader values are invalid for serialization.