mrrc

Module mods

MODS (Metadata Object Description Schema) serialization of MARC records.

This module provides conversion of MARC records to MODS XML format, a metadata schema developed by the Library of Congress that is richer than Dublin Core and suitable for detailed bibliographic description.

MODS includes elements for:

Titles (with type information)

Names (personal, corporate, conference)

Identifiers (ISBN, ISSN, etc.)

Language

Physical description

Subject headings with authority information

Locations and holdings information

Related resources

Examples

use mrrc::{Record, Field, Leader, mods};

let mut record = Record::new(Leader::default());
let mut field = Field::new("245".to_string(), '1', '0');
field.add_subfield('a', "Title".to_string());
record.add_field(field);

let mods_xml = mods::record_to_mods_xml(&record)?;
println!("{}", mods_xml);

Functions

record_to_mods_xml

Convert a MARC record to MODS XML format.